\name{expit.fn}
\alias{expit.fn}

\title{Expit function}

\description{

  This is the inverse of the logit function.
  
}

\usage{

expit.fn(x)


}

\arguments{

  \item{x}{any scalar, vector or matrix x}
    
}

\details{

  The function is mainly used internally in \code{clustglm}.
  
}

\value{

  A real-valued scalar, vector or matrix.
  
}

